+++ google-cloud/container-linux/kubernetes/outputs.tf
@@ -13,3 +13,7 @@ output "network_name" {
 output "network_self_link" {
   value = "${google_compute_network.network.self_link}"
 }
+
+output "kubeconfig" {
+  value = "${module.bootkube.kubeconfig}"
+}